@@ -1,16 +1,14 @@
-﻿'use strict';
-
-var njr = require('./core'),
-  nj = require('nornj'),
-  ReactDOM = require('react-dom');
+﻿import njr from './core';
+import nj from 'nornj';
+import ReactDOM from 'react-dom';
 
 //渲染内联标签组件
-function renderTmplTag(data, selector, isAuto) {
+export function renderTmplTag(data, selector, isAuto) {
   var tags = getTmplTag(selector, isAuto),
     ret = [];
 
-  nj.each(tags, function (tag) {
-    var tmpl = nj.compileH(tag.innerHTML, tag.id),
+  nj.each(tags, (tag) => {
+    const tmpl = nj.compileH(tag.innerHTML, tag.id),
       parentNode = tag.parentNode;
 
     ret.push(ReactDOM.render(tmpl(data), parentNode));
@@ -20,7 +18,7 @@ function renderTmplTag(data, selector, isAuto) {
 }
 
 //获取全部内联组件
-function getTmplTag(selector, isAuto) {
+export function getTmplTag(selector, isAuto) {
   if (!selector) {
     selector = 'script[type="text/nornj"]' + (isAuto ? '[autorender]' : '');
   }
@@ -29,12 +27,6 @@ function getTmplTag(selector, isAuto) {
 }
 
 //Set initial data for inline component
-function setInitialData(data) {
+export function setInitialData(data) {
   njr.initialData = data;
 }
-
-module.exports = {
-  renderTmplTag: renderTmplTag,
-  getTmplTag: getTmplTag,
-  setInitialData: setInitialData
-};
