@@ -1,7 +1,6 @@
-﻿'use strict';
+﻿import nj from 'nornj';
 
-var nj = require('nornj');
-
-nj.registerExpr('brace', function(options) {
-  return '{' + options.result() + '}';
-}, { newContext: false, useString: false });
+nj.registerExpr('brace', (options) => '{' + options.result() + '}', {
+  newContext: false,
+  useString: false
+});
